% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Data}
\alias{Data}
\alias{ADAE}
\alias{ADSL}
\title{Example dataset}
\format{An object of class \code{data.frame} with 3187 rows and 3 columns.}
\usage{
data(ADAE)

data(ADSL)
}
\description{
Randomly generated datasets used for running examples in
this package.
}
\details{
\code{ADAE} is the "adverse event analysis dataset", it contains four columns:\cr
\code{USUBJID}: unique subject id \cr
\code{AEBODSYS}: SoC of the AE \cr
\code{AEDECOD}: PT of the AE \cr


\code{ADSL} is the "subject level analysis dataset", it contains three columns: \cr
\code{USUBJID}: unique subject id \cr
\code{TREATMENT}: treatment applied on each subject, subjects are divided into control or treatment group based on this
\code{TRTCTR}: indicitator for treatment group (TRTCTR=1, TREATMENT = "xyz" or 'xyz2') or
control group (TRTCTR=0, TREATMENT='control' or 'control2')
}
\keyword{datasets}
