% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierIsingplot.R
\name{HierIsingplot}
\alias{HierIsingplot}
\alias{gci3}
\alias{HIPLOT}
\title{Plot for two Bayesian models}
\usage{
gci3(aedata, ptnum = 10, conf.level = 0.95)

HIPLOT(hierdata, isingdata, aedata, ptnum = 10,
  param = "risk difference")
}
\arguments{
\item{ptnum}{positive integer, number of AEs to be selected or plotted, default is 10}

\item{hierdata}{output from function \code{\link{Hier}}}

\item{isingdata}{output from function \code{\link{Ising}}}

\item{param}{a string, either "odds ratio" or "risk difference", indicate which summary statistic to be based on to plot the top AEs,
default is "risk difference"}
}
\description{
These two functions are for plotting the top AEs selected by Bayesian hierarchical model
and Bayesian model with Ising prior.
}
\details{
\code{gci3} calculates the incidence rate for each adverse event (AE) in treatment
and in control group and also the confidence interval for this incidence getting from fisher exact test
then the PT of AEs with the highest \code{ptnum} incidence rate difference between treatment and control
group are returned. \cr
\code{HIPLOT} first selects the top \code{ptnum} (an integer) AE based on the selected statistic (either "odds ratio" or "risk difference").
Then it plots the mean, 2.5% quantile, 97.5% quantile of the selected statistic of these AE from both two models(methods). It seperates
the AEs slected by both Bayesian methods from AEs selected by only one method. Also it indicates whether the AE selected by these two Bayesian models were also selected by only based on
incidence difference (function \code{gci3}). \cr
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)

# run the Hierarchical model
INITS1<-list(mu.gamma.0=0.1, tau.gamma.0=0.1, mu.theta.0=0.1, tau.theta.0=0.1, alpha.pi=2, beta.pi=2)
INITS2<-list(mu.gamma.0=1, tau.gamma.0=1, mu.theta.0=1, tau.theta.0=1, alpha.pi=10, beta.pi=10)
INITS <- list(INITS1,INITS2)
HIERDATA<-Hier(aedata=AEdata, inits=INITS, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=2)

# run the Ising model
RHO<-rep(1,dim(AEdata)[1])
THETA<-0.02
SIM<-c(5000,1000,20)
BETA.AB<-c(0.25, 0.75)
ISINGDATA<-Ising(aedata = AEdata, beta.ab = BETA.AB, rho = RHO, theta = THETA, sim = SIM)

gci3(aedata=AEdata)
HIPLOT(hierdata=HIERDATA, isingdata=ISINGDATA, aedata=AEdata)
HIPLOT(hierdata=HIERDATA, isingdata=ISINGDATA, aedata=AEdata, ptnum=15, param="odds ratio")
}



}
\seealso{
\code{\link{preprocess}}, \code{\link{Hier}}, \code{\link{Ising}}
}
