% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hiermodel.R
\name{Hiermodel}
\alias{Hiermodel}
\alias{Hier_history}
\alias{sum_Hier}
\alias{Hier}
\alias{Hiergetpi}
\alias{Hierplot}
\alias{Hiertable}
\title{Three stages Bayesian Hierarchical Model}
\usage{
Hier_history(aedata, n_burn, n_iter, thin, n_adapt, n_chain,
  alpha.gamma = 3, beta.gamma = 1, alpha.theta = 3, beta.theta = 1,
  mu.gamma.0.0 = 0, tau.gamma.0.0 = 0.1, alpha.gamma.0.0 = 3,
  beta.gamma.0.0 = 1, lambda.alpha = 0.1, lambda.beta = 0.1,
  mu.theta.0.0 = 0, tau.theta.0.0 = 0.1, alpha.theta.0.0 = 3,
  beta.theta.0.0 = 1)

sum_Hier(hierraw)

Hier(aedata, n_burn, n_iter, thin, n_adapt, n_chain, alpha.gamma = 3,
  beta.gamma = 1, alpha.theta = 3, beta.theta = 1,
  mu.gamma.0.0 = 0, tau.gamma.0.0 = 0.1, alpha.gamma.0.0 = 3,
  beta.gamma.0.0 = 1, lambda.alpha = 0.1, lambda.beta = 0.1,
  mu.theta.0.0 = 0, tau.theta.0.0 = 0.1, alpha.theta.0.0 = 3,
  beta.theta.0.0 = 1)

Hiergetpi(aedata, hierraw)

Hierplot(hierdata, ptnum = 10, param = "risk difference",
  OR_ylim = c(0, 5))

Hiertable(hierdata, ptnum = 10, param = "risk difference")
}
\arguments{
\item{aedata}{output from function \code{\link{preprocess}}}

\item{n_burn}{integer, number of interations without saving posterior samples}

\item{n_iter}{integer, number of interations saving posterior samples with every \code{thin}th iteration}

\item{thin}{integer, samples are saved for every \code{thin}th iteration}

\item{n_adapt}{integer, number of adaptations}

\item{n_chain}{number of MCMC chains}

\item{hierdata}{utput from function \code{\link{Hier}}}

\item{ptnum}{positive integer, number of AEs to be selected or plotted, default is 10}

\item{param}{a string, either "odds ratio" or "risk difference", indicate which summary statistic to be based on to plot the top AEs,
default is "risk difference"}

\item{OR_ylim}{a numeric vector of two elements, used to set y-axis limit for plotting based on "odds ratio"}

\item{alpha.gamma...}{alpha.gamma and other parameters are hyperparameters for prior distribution,
see the reference for the meaning of each parameters}
}
\value{
\strong{\code{Hier_history}}\cr
It returns a dataframe with each column represent one parameter
and each row is the output from one sampling/one iteraction (like the output of \code{\link[R2jags]{coda.samples}}) \cr
\strong{\code{sum_Hier}} \cr
It returns the summary statistics for each parameter recorded by \code{Hier_history}. \cr
\strong{\code{Hier}} \cr
It returns the summary statistics for each AE, combining with raw data.
The summary statistics including:
summary statistics for incidence rate difference (mean, 2.5\% and 97.5\% percentile);
summary statistics for odds ratio (mean, 2.5\% and 97.5\% percentile).
The other columns include SoC, PT, Nt, Nc, AEt, and AEc.
\strong{\code{Hiergetpi}}: \cr
This function calculates pit (incidence of AE in treatment group) and
pic (incidence of AE in control group) from the output of \code{Hier_history}.\cr
}
\description{
Functions here are to take the orginized data (output from
  \code{preprocess}) and apply the three stages Bayesian Hierarchical Model.
  See details for model description and difference between each function.
}
\details{
\strong{Model}: \cr Here the 3-stage hierarchical bayesian model was
  used to model the probability of AEs. It is model 1b (Bayesian Logistic
  Regression Model with Mixture Prior on Log-OR) in H. Amy Xia , Haijun Ma &
  Bradley P. Carlin (2011) Bayesian Hierarchical Modeling for Detecting
  SaBCIy Signals in Clinical Trials, Journal of Biopharmaceutical Statistics,
  21:5, 1006-1029, DOI: 10.1080/10543406.2010.520181) \cr
  \strong{\code{Hier_history}}: \cr
  This function takes formatted Binomial data and
  output Gibbs sample of the defined parametes. The output is a dataframe
  with each column represent one parameter and each row is the output from
  one sampling/one iteration. Diff, OR, gamma, and theta are the parameters recorded.\cr
  \emph{Diff}: is the difference of incidence
  of AE between treatment and control group (treatment - control) \cr
  \emph{OR}: is the odds ratio for the incidence of AE between treatment and
  control group (treatment over control): t(1-c)/c(1-t), where t,c are the incidences of one AE for treatment
  and control group, respectively.\cr
  \emph{gamma}: logit(incidence of AE in control group) = gamma \cr
  \emph{theta}: logit(incidence of AE in treatment group) = gamma + theta; and OR = exp(theta) \cr
  The result for Diff, OR, gamma, and theta are ordered by j and then by b.
  For example the result is like Diff.1.1, Diff.2.1, Diff.3.1, Diff.4.1 Diff.1.2, Diff.2.2,
  Diff.3.2 and so on \cr
  \strong{\code{sum_Hier}}:\cr
  This function takes the output from \code{Hier_history} and return the summary
  statistics for each parameter recorded by \code{Hier_history}. The summary function is applied on each column. \cr
  \strong{\code{Hier}}:\cr
  This function takes the same input as \code{Hier_history} and calculate the summary statistics
  for output from \code{Hier_history}.
  It outputs the summary statistics for each AE, combining with raw data. \cr
  \strong{\code{Hiergetpi}}: \cr
  This function calculates pit (incidence of AE in treatment group) and
  pic (incidence of AE in control group) from the output of \code{Hier_history}
  The output is used for Loss function. \cr
  \code{Hierplot} first selects the top \code{ptnum} (an integer) AE based on the selected statistic (either "odds ratio" or "risk difference").
  Then it plots the mean, 2.5% quantile, 97.5% quantile of the selected statistic of these AE. It shows the PTs of these AE from same SOC in same
  color.  \cr
  \code{Isingtable} creates a table for the detailed information for AE plotted in \code{Hierplot}.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)
HIERRAW<-Hier_history(aedata=AEdata, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=2)
HIERRAW2<-Hier_history(aedata=AEdata, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=1)
# you can specify the hyperparameter as shown below
HIERRAW3<-Hier_history(aedata=AEdata, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=1,
alpha.gamma=5, beta.gamma=1, alpha.theta=3, beta.theta=1, mu.gamma.0.0=0.1, tau.gamma.0.0=0.1, alpha.gamma.0.0=5,
beta.gamma.0.0=1, lambda.alpha=0.1, lambda.beta=0.1, mu.theta.0.0=0.1, tau.theta.0.0=0.1,alpha.theta.0.0=3, beta.theta.0.0=1)

HIERDATA<-Hier(aedata=AEdata, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=2)
HIERPI<-Hiergetpi(aedata=AEdata, hierraw=HIERRAW)

Hierplot(HIERDATA)
Hierplot(HIERDATA, ptnum=15, param="odds ratio")
HIERTABLE<-Hiertable(HIERDATA)
HIERTABLE2<-Hiertable(HIERDATA, ptnum=15, param="odds ratio")
}

}
\references{
H. Amy Xia , Haijun Ma & Bradley P. Carlin (2011) Bayesian
  Hierarchical Modeling for Detecting Safety Signals in Clinical Trials,
  Journal of Biopharmaceutical Statistics, 21:5, 1006-1029, DOI:
  10.1080/10543406.2010.520181)
}
\seealso{
\code{\link{preprocess}}
}
