% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isingmodel.R
\name{Isingpriormodel}
\alias{Isingpriormodel}
\alias{Ising_history}
\alias{sum_Ising}
\alias{Ising}
\alias{Isinggetpi}
\alias{Isingplot}
\alias{Isingtable}
\title{Bayesian Model with Ising Latent Variables}
\usage{
Ising_history(aedata, n_burn, n_iter, thin, alpha_ = 0.25,
  beta_ = 0.75, alpha.t = 0.25, beta.t = 0.75, alpha.c = 0.25,
  beta.c = 0.75, rho, theta)

sum_Ising(isingraw)

Ising(aedata, n_burn, n_iter, thin, alpha_ = 0.25, beta_ = 0.75,
  alpha.t = 0.25, beta.t = 0.75, alpha.c = 0.25, beta.c = 0.75,
  rho, theta)

Isinggetpi(aedata, isingraw)

Isingplot(isingdata, ptnum = 10, param = "risk difference",
  OR_xlim = c(0, 5))

Isingtable(isingdata, ptnum = 10, param = "risk difference")
}
\arguments{
\item{aedata}{output from function \code{\link{preprocess}}}

\item{n_burn}{number of burn in for Gibbs Sampling}

\item{n_iter}{number of interation for Gibbs Sampling}

\item{thin}{thin for Gibbs Samping, parameters are recorded every thin-th interation}

\item{alpha_}{numeric, is the prior for beta distribution, beta distribution for both treatment and control group, alpha parameter of beta distribution}

\item{beta_}{numeric, is the prior for beta distribution, beta distribution for both treatment and control group, beta parameter of beta distribution}

\item{alpha.t}{numeric, is the prior for beta distribution, beta distribution for treatment group, alpha parameter of beta distribution}

\item{beta.t}{numeric, is the prior for beta distribution, beta distribution for  treatment group, beta parameter of beta distribution}

\item{alpha.c}{numeric, is the prior for beta distribution, beta distribution for control group, alpha parameter of beta distribution}

\item{beta.c}{numeric, is the prior for beta distribution, beta distribution for control group, beta parameter of beta distribution}

\item{rho}{either a number or numeric vector with length equals to the number of rows of data frame aedata.
If it is a single number, then all adverse events use the same hyperparameter of rho. If it is a numeric vector, then each AE has
its own hyperparameter of rho, and the sequence of rho value for each AE should be the same as the sequence of AE in aedata (AE in
aedata should be ordered by b and j).}

\item{theta}{numeric, \code{rho} and \code{theta} are parameters for Ising prior}

\item{isingraw}{output from function \code{\link{Ising_history}}}

\item{isingdata}{output from function \code{\link{Ising}}}

\item{ptnum}{positive integer, number of AEs to be selected or plotted, default is 10}

\item{param}{a string, either "odds ratio" or "risk difference", indicate which summary statistic to be based on to plot the top AEs,
default is "risk difference"}

\item{OR_ylim}{a numeric vector of two elements, used to set y-axis limit for plotting based on "odds ratio"}
}
\value{
\code{Ising_history} returns a list with 3 matries, gamma, pi.t, and pi.c.
Each row of these three matries correspond to one recoreded iteration.
And each column of these three matries correspond to each of the AE.\cr
\emph{gamma} matrix records the indicator of non-differential risk for each AE. \cr
\emph{pi.t} records the incidence risk of treatment group for each AE. \cr
\emph{pi.c} records the incidence risk of control group for each AE. \cr

\code{sum_Ising} returns the summary statistics for each AE.(Note: all the apply function are apply
on column, thus will result the summary statistics for each AE). \cr

\code{Ising} returns the summary statistics for each AE combining with the raw data. The summary statistics include
summary statistics for incidence rate difference (mean, 2.5\% and 97.5\% percentile),
summary statistics for odds ratio (mean, 2.5\% and 97.5\% percentile), and also the mean of gamma, with is the probability of gamma=1.
The other columns from raw data include SoC, PT, Nt, Nc, AEt, AEc. \cr

\strong{\code{Isinggetpi}}: \cr
This function calculates pit (incidence of AE in treatment group) and
pic (incidence of AE in control group) from the output of \code{Ising_history}. \cr
}
\description{
Functions here are to take the orginized data (output from
  \code{preprocess}) and apply the Bayesian model with Ising latent variables.
  See details for model description and difference between each function.
}
\details{
\strong{Model}:\cr
The model is based on the paper McEvoy, Bradley W., Rajesh R. Nandy, and Ram C. Tiwari.
"Bayesian approach for clinical trial safety data using an Ising prior."
Biometrics 69.3 (2013): 661-672. \cr
\strong{\code{Ising_history}}:\cr
This function takes formatted Binomial data and
performs the Bayesian analysis with Ising prior.#'
\strong{\code{sum_Ising}}:\cr
This function takes the result from function \code{Ising_history} as input
and summarizes the result from \code{Ising_history}.\cr
\strong{\code{Ising}}:\cr
This function takes the same parameters as function \code{Ising_history}
this function will take first take the result from \code{Ising_history} and
then summarize the result for each AE and merge the result with the raw data,and also the Raw risk difference, Raw odds ratio calculated from raw data.
\strong{\code{Isinggetpi}}: \cr
This function calculates pit (incidence of AE in treatment group) and
pic (incidence of AE in control group) from the output of \code{Ising_history}
\code{Isingplot} first selects the top \code{ptnum} (an integer) AE based on the selected statistic (either "odds ratio" or "risk difference").
Then it plots the mean, 2.5% quantile, 97.5% quantile of the selected statistic of these AE. It shows the PTs of these AE from same SOC in same
color.  \cr
\code{Isingtable} creates a table for the detailed information for AE plotted in \code{Isingplot}.
}
\note{
The incidence difference is calculated by (t-c) and the odds ratio is calculated by t(1-c)/c(1-t), where t,c are
the incidences of one AE for treatment and control group, respectively.\cr
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)
RHO<-rep(1,dim(AEdata)[1])
ISINGRAW<-Ising_history(aedata = AEdata, n_burn=1000, n_iter=5000, thin=20, rho=1, theta=0.02)
ISINGRAW2<-Ising_history(aedata = AEdata, n_burn=1000, n_iter=5000, thin=20, alpha_=0.5, beta_=0.5,
                           alpha.t=0.5, beta.t=0.5, alpha.c=0.25, beta.c=0.75, rho=RHO, theta=0.02)
SUM_ISING<-sum_Ising(ISINGRAW)
ISINGDATA<-Ising(aedata = AEdata, n_burn=1000, n_iter=5000, thin=20, rho=1, theta=0.02)
ISINGDATA<-Ising(aedata = AEdata, n_burn=1000, n_iter=5000, thin=20, alpha_=0.5, beta_=0.5,
                           alpha.t=0.5, beta.t=0.5, alpha.c=0.25, beta.c=0.75, rho=RHO, theta=0.02)
ISINGPI<-Isinggetpi(aedata = AEdata, isingraw=ISINGRAW)

Isingplot(ISINGDATA)
Isingplot(ISINGDATA, ptnum=15, param="odds ratio", OR_ylim=c(1,10))
ISINGTABLE<-Isingtable(ISINGDATA)
ISINGTABLE2<-Isingtable(ISINGDATA, ptnum=15, param="odds ratio")
}

}
\references{
McEvoy, Bradley W., Rajesh R. Nandy, and Ram C. Tiwari.
"Bayesian approach for clinical trial safety data using an Ising prior."
Biometrics 69.3 (2013): 661-672.
}
\seealso{
\code{\link{preprocess}}
}
