% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Raw data process}
\usage{
preprocess(adsl, adae)
}
\arguments{
\item{adsl}{subject level analysis dataset, it is a .csv file, it has to contain at least two columns, "USUBJID" and "TRTCTR", "TRTCTR"
is the indicator for treatment and control group. TRTCTR=1 for treatment group and TRTCTR=0 for control group.}

\item{adae}{adverse event analysis dataset, it is a .csv file, it has to contain at least three columns, "USUBJID", "AEBODSYS",and "AEDECOD"}
}
\value{
\code{preprocess} returns a dataframe with following columnss:\cr
\emph{AEBODSYS}: Body System or SoC for AE \cr
\emph{AEDECOD}: Peferred Term (PT) for AE \cr
\emph{AEc}: number of patients in control group having each specific AE \cr
\emph{AEt}: number of patients in treatment group having each specific AE \cr
\emph{Nc}: number of total patients in control group \cr
\emph{Nt}: number of total patients in treatmetn group \cr
\emph{b}: integer represents each Soc \cr
\emph{i}: integer represents each PT \cr
\emph{j}: order of PT in each SoC \cr
}
\description{
\code{preprocess} reorganizes the datasets by adverse events (AE), so that it can be used for further analysis.\cr
}
\details{
\code{preprocess} sums up the number of patients in control and treatment group.
Also this function calculates the number of patients
experiencing the AE for each AE in control and treatment group. It also contains columns, "b", "i", "j" for further analysis.
}
\note{
Make sure that adsl and adae contain the required columns with exact the same column names as listed in "parameters" section.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
preprocess(adsl=ADSL, adae=ADAE)
}

}
