% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdataprocess.R
\name{rawdataprocess}
\alias{rawdataprocess}
\alias{preprocess}
\alias{preprocess2}
\alias{preprocess3}
\alias{preprocess4}
\title{Raw data process}
\usage{
preprocess(adsl, adae, TreatCol, drug)

preprocess2(adsl, adae, TreatCol, drug)

preprocess3(adsl, adae, TreatCol, treatcode, controlcode)

preprocess4(adsl, adae, TreatCol, treatcode, controlcode)
}
\arguments{
\item{adsl}{subject level analysis dataset, it is a .sas7bdat file}

\item{adae}{adverse event analysis dataset, it is a .sas7bdat file}

\item{TreatCol}{a string}

\item{drug}{a string}

\item{treatcode}{a vector of strings}

\item{controlcode}{a vector of strings}
}
\value{
\code{preprocess} returns a dataframe with following columns:\cr
\emph{USUBJID}: the unique subject id for all subjects in the trail with or without AE \cr
\emph{TreatCol}: the column based on which the treatment/control division is performed \cr
\emph{TreatCol}: the column based on which the treatment/control division is performed\cr
\emph{AEBODSYS}: Body System or SoC for AE\cr
\emph{AEDECOD}: Peferred Term (PT) for AE (if a subject did not have any AE, AEBODSYS and AEDECOD are NA)\cr
\emph{trt}: indication of treatment group (1) or control group (0)\cr

\code{preprocess2} returns a dataframe with following columnss:\cr
\emph{AEBODSYS}: Body System or SoC for AE \cr
\emph{AEDECOD}: Peferred Term (PT) for AE \cr
\emph{AEc}: number of patients in control group having each specific AE \cr
\emph{AEt}: number of patients in treatment group having each specific AE \cr
\emph{Nc}: number of total patients in control group \cr
\emph{Nt}: number of total patients in treatmetn group \cr
\emph{b}: integer represents each Soc \cr
\emph{i}: integer represents each PT \cr
\emph{j}: order of PT in each SoC \cr
}
\description{
\code{preprocess} divides the subjects(patients) in adsl into two groups
by creating a new column 'trt' with trt=1 for treatment group and trt=0 for control group.
Also this function only takes the columns related to further analysis. \cr
\code{preprocess2} first execute function \code{preprocess}
and reorganize the dataset by adverse events (AE), so that it can be used for further analysis.\cr

\code{preprocess3} is same as \code{preprocess} except that \code{preprocess3} takes 'treatcode' and 'controlcode' as
input instead of 'drug'. Like 'drug' in \code{preprocess}, 'treatcode' and 'controlcode' here are also used to determine
whether one subject is belong to treatment or control group. If the value under column TreatCol is
an element of 'treatcode' then this subject belongs to treatment group (trt=1), if the value under column TreatCol
is an element of 'controlcode' then this subject belogns to control group (trt=0). For subject with value
under column TreatCol not an element of either 'treatcode' or 'controlcode', subject is removed from the dataset.
control group (trt=0). \cr

\code{preprocess4} is same as \code{preprocess2} except it first exectue function \code{preprocess3}. \cr

\code{preprocess3} and \code{preprocess4} provide more flexiblity for processing the raw data.
}
\details{
Parameter "TreatCol" is the column based on which the division is performed,
paramter "drug" is to determine the value of trt,
if the value under TreatCol contains drug then trt=1, otherwise trt=0.\cr
\code{preprocess2} sums up the number of patients in control and treatment group.
Also this function calculates the number of patients in control and treatment group
experiencing the AE for each AE. It also contains columns, "b", "i", "j" for further analysis.
}
\note{
Both functions only select the patients that at least took one dose of the treatment
and only select the adverse events occurs within 30 days after taking the first dose. \cr
\code{preprocess} and \code{preprocess3}summarize the information about AE, and its output can be used
for exploratory analysis. \cr
The outout of \code{preprocess2} and \code{preprocess4} are organized by column b and column j for further analysis.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
preprocess(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz")
preprocess2(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz")
preprocess3(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", treatcode=c("xyz", "xyz2"), controlcode=c("control", "control2))
preprocess4(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", treatcode=c("xyz", "xyz2"), controlcode=c("control", "control2))

}

}
