% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherETplot.R
\name{FisherETplot}
\alias{FisherETplot}
\alias{FETplot}
\alias{FETtable}
\title{Plot based on Fisher exact test}
\usage{
FETplot(aedata, ptnum = 10, conf.level = 0.95)

FETtable(aedata, ptnum = 10, conf.level = 0.95)
}
\arguments{
\item{aedata}{output from function \code{\link{preprocess}}}

\item{ptnum}{positive integer, number of AEs to be plotted, default is 10}

\item{conf.level}{number between 0 and 1, confidence level for constructing
confidence interval, default is 0.95}
}
\description{
Both functions \code{FETplot} plot the top AEs with
top ptnum (one parameter, an integer) highest difference of incidence rate
between treatment and control group(treatment - control). The incidence and
confidence interval from treatment and control group are plotted. AEs from
the same SoC is plotted in same color. Function \code{FETtable} return a table containing
information about AE selected.
}
\details{
The incidence and confidence interval for both control and treatment
group are calculated by Fisher exact test. \code{FETplot} plots out the AEs with confidence
interval, while \code{FETtable} output a table containig the information of
AEs in the plot. The table from \code{FETtable} has the following columns:\
AEBODSYS: SOC of the AE
AEDECOD: PT term of the AE
AEc: number of occurence of the AE in control group
AEt: number of occurence of the AE in treatment group
Nc: number of subjects in control group
Nt: number of subjects in treatment group
c_mean: incidence of the AE in control group
c_lowerbd: lower bound of the confidence interval of incidence the AE in control group
c_upperbd: upper bound of the confidence interval of incidence the AE in control group
t_mean: incidence of the AE in treatment group
t_lowerbd: lower bound of the confidence interval of incidence the AE in treatment group
t_upperbd: upper bound of the confidence interval of incidence the AE in treatment group
mean_diff: difference of incidence of the AE in treatment group and control group
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)
FETplot(aedata=AEdata)
FETplot(aedata=AEdata, ptnum=15, conf.level=0.9)
FETtable(aedata=AEdata)
FETtable(aedata=AEdata, ptnum=15, conf.level=0.9)
}
}
\seealso{
\code{\link{preprocess}}
}
