% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherETplot.R
\name{FisherETplot}
\alias{FisherETplot}
\alias{gci}
\alias{gci2}
\title{Plot based on Fisher exact test}
\usage{
gci(aedata, ptnum = 10, conf.level = 0.95)

gci2(adsl, adae, TreatCol, drug, ptnum = 10, conf.level = 0.95)
}
\arguments{
\item{aedata}{output from function \code{\link{preprocess2}}}

\item{ptnum}{positive integer, number of AEs to be plotted, default is 10}

\item{conf.level}{number between 0 and 1, confidence level for constructing
confidence interval, default is 0.95}

\item{adsl}{subject level analysis dataset, it is a .sas7bdat file}

\item{adae}{adverse event analysis dataset, it is a .sas7bdat file}

\item{TreatCol}{a string}

\item{drug}{a string}
}
\description{
Both functions \code{gci} and \code{gci2} plot the top AEs with
top ptnum (one parameter, an integer) highest difference of incidence rate
between treatment and control group(treatment - control). The incidence and
confidence interval from treatment and control group are plotted. AEs from
the same SoC is plotted in same color.
}
\details{
The incidence and confidence interval for both control and treatment
group are calculated by Fisher exact test.\cr The only difference between
\code{gci} and \code{gci2} is that \code{gci} takes the output of
\code{preprocess} as the input, while \code{gci2} takes the raw data as the
input.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess2(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz")
gci(aedata=AEdata)
gci(aedata=AEdata, ptnum=15, conf.level=0.9)
gci2(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz")
gci2(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz", ptnum=15, conf.level=0.9)
}
}
\seealso{
\code{\link{preprocess2}}
}
