% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierIsingplot.R
\name{HierIsingplot}
\alias{HierIsingplot}
\alias{HIplot}
\alias{HItable}
\title{Plot for two Bayesian models}
\usage{
HIplot(hierdata, isingdata, aedata, ptnum = 10,
  param = "risk difference", OR_ylim = c(0, 5))

HItable(hierdata, isingdata, aedata, ptnum = 10,
  param = "risk difference")
}
\arguments{
\item{hierdata}{output from function \code{\link{Hier}}}

\item{isingdata}{output from function \code{\link{Ising}}}

\item{ptnum}{positive integer, number of AEs to be selected or plotted, default is 10}

\item{param}{a string, either "odds ratio" or "risk difference", indicate which summary statistic to be based on to plot the top AEs,
default is "risk difference"}

\item{OR_ylim}{a numeric vector of two elements, used to set y-axis limit for plotting based on "odds ratio"}
}
\value{
\code{HIplot} returns a plot for the top \code{ptnum} (an integer) AE based on the selected statistics (either "odds ratio" or "risk difference"). Mean, 2.5%
quantile, 97.5% quantile of the selected statistics of these AE from both two models were plotted.
\cr
\code{HItable} returns a table for the detailed information for AE plotted in \code{HIplot}. It contains a new column, "rank_diff_mean" or "rank_OR_mean"
(based on the \code{param}), besides the columns of output from \code{\link{Hier}} or \code{\link{Ising}}. This new column is the rank of "Diff_mean" or "OR_mean"
of the AE in each method.
}
\description{
These two functions are for plotting the top AEs selected by Bayesian hierarchical model
and Bayesian model with Ising prior and producing table with detailed information fo these AEs.
}
\details{
\code{HIplot} first selects the top \code{ptnum} (an integer) AE based on the selected statistic (either "odds ratio" or "risk difference").
Then it plots the mean, 2.5% quantile, 97.5% quantile of the selected statistic of these AE from both two models(methods). It seperates
the AEs slected by both Bayesian methods from AEs selected by only one method. Also it indicates whether the AE selected by these two Bayesian models were also selected by only based on
incidence difference (function \code{\link{BCItable}}). \cr
\code{HItable} creates a table for the detailed information for AE plotted in \code{HIplot}.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)

# run the Hierarchical model
INITS1<-list(mu.gamma.0=0.1, tau.gamma.0=0.1, mu.theta.0=0.1, tau.theta.0=0.1, alpha.pi=2, beta.pi=2)
INITS2<-list(mu.gamma.0=1, tau.gamma.0=1, mu.theta.0=1, tau.theta.0=1, alpha.pi=10, beta.pi=10)
INITS <- list(INITS1,INITS2)
HIERDATA<-Hier(aedata=AEdata, inits=INITS, n_burn=1000, n_iter=1000, thin=20, n_adapt=1000, n_chain=2)

# run the Ising model
RHO<-rep(1,dim(AEdata)[1])
THETA<-0.02
SIM<-c(5000,1000,20)
BETA.AB<-c(0.25, 0.75)
ISINGDATA<-Ising(aedata = AEdata, beta.ab = BETA.AB, rho = RHO, theta = THETA, sim = SIM)

HIplot(hierdata=HIERDATA, isingdata=ISINGDATA, aedata=AEdata)
HIplot(hierdata=HIERDATA, isingdata=ISINGDATA, aedata=AEdata, ptnum=15, param="odds ratio")

HItable(hierdata=HIERDATA, isingdata=ISINGDATA)
HItable(hierdata=HIERDATA, isingdata=ISINGDATA, ptnum=15, param="odds ratio")

}

}
\seealso{
\code{\link{preprocess}}, \code{\link{Hier}}, \code{\link{Ising}}, \code{\link{BCItable}}
}
