% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isingmodel.R
\name{Isingpriormodel}
\alias{Isingpriormodel}
\alias{Ising_history}
\alias{sum_Ising}
\alias{Ising}
\alias{Isinggetpi}
\title{Bayesian Model with Ising Latent Variables}
\usage{
Ising_history(aedata, beta.ab, rho, theta, sim)

sum_Ising(isingraw)

Ising(aedata, beta.ab, rho, theta, sim)

Isinggetpi(aedata, isingraw)
}
\arguments{
\item{aedata}{output from function \code{\link{preprocess2}}}

\item{beta.ab}{numeric vector with length of 2, is the prior for beta distribution}

\item{rho}{numeric vector with length equals to the number of rows of data frame aedata}

\item{theta}{numeric, \code{rho} and \code{theta} are parameters for Ising prior}

\item{sim}{numeric vecotr with length of 3, integer for each element, sim[1] is the
number of iterations of brun in; sim[2] is the number of interactions to recorded;
sim[3] is like the parameter thin in MCMC settings. The total number of iterations running is
\eqn{sim[1]+sim[2]*sim[3]}}

\item{isingraw}{output from function \code{\link{Ising_history}}}
}
\value{
\code{Ising_history} returns a list with 3 matries, gamma, pi.t, and pi.c.
Each row of these three matries correspond to one recoreded iteration.
And each column of these three matries correspond to each of the AE.\cr
\emph{gamma} matrix records the indicator of non-differential risk for each AE. \cr
\emph{pi.t} records the incidence risk of treatment group for each AE. \cr
\emph{pi.c} records the incidence risk of control group for each AE. \cr

\code{sum_Ising} returns the summary statistics for each AE.(Note: all the apply function are apply
on column, thus will result the summary statistics for each AE). \cr

\code{Ising} returns the summary statistics for each AE combining with the raw data. The summary statistics include
summary statistics for incidence rate difference (mean, 2.5\% and 97.5\% percentile),
summary statistics for odds ratio (mean, 2.5\% and 97.5\% percentile), and also the mean of gamma, with is the probability of gamma=1.
The other columns from raw data include SoC, PT, Nt, Nc, AEt, AEc. \cr

\strong{\code{Isinggetpi}}: \cr
This function calculates pit (incidence of AE in treatment group) and
pic (incidence of AE in control group) from the output of \code{Ising_history}. \cr
}
\description{
Functions here are to take the orginized data (output from
  \code{preprocess2}) and apply the Bayesian model with Ising latent variables.
  See details for model description and difference between each function.
}
\details{
\strong{Model}:\cr
The model is based on the paper McEvoy, Bradley W., Rajesh R. Nandy, and Ram C. Tiwari.
"Bayesian approach for clinical trial safety data using an Ising prior."
Biometrics 69.3 (2013): 661-672. \cr
\strong{\code{Ising_history}}:\cr
This function takes formatted Binomial data and
performs the Bayesian analysis with Ising prior.#'
\strong{\code{sum_Ising}}:\cr
This function takes the result from function \code{Ising_history} as input
and summarizes the result from \code{Ising_history}.\cr
\strong{\code{Ising}}:\cr
This function takes the same parameters as function \code{Ising_history}
this function will take first take the result from \code{Ising_history} and
then summarize the result for each AE and merge the result with the raw data.
\strong{\code{Isinggetpi}}: \cr
This function calculates pit (incidence of AE in treatment group) and
pic (incidence of AE in control group) from the output of \code{Ising_history}
}
\note{
The incidence difference is calculated by (t-c) and the odds ratio is calculated by t(1-c)/c(1-t), where t,c are
the incidences of one AE for treatment and control group, respectively.\cr
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess2(adsl=ADSL, adae=ADAE, TreatCol="TREATMENT", drug="xyz")
RHO<-rep(1,dim(AEdata)[1])
THETA<-0.02
SIM<-c(5000,1000,20)
BETA.AB<-c(0.25, 0.75)
ISINGRAW<-Ising_history(aedata = AEdata, beta.ab = BETA.AB, rho = RHO, theta = THETA, sim = SIM)
SUM_ISING<-sum_Ising(ISINGRAW)
ISINGMODEL<-Ising(aedata = AEdata, beta.ab = BETA.AB, rho = RHO, theta = THETA, sim = SIM)
ISINGPI<-Isinggetpi(aedata = AEdata, isingraw=ISINGRAW)
}

}
\references{
McEvoy, Bradley W., Rajesh R. Nandy, and Ram C. Tiwari.
"Bayesian approach for clinical trial safety data using an Ising prior."
Biometrics 69.3 (2013): 661-672.
}
\seealso{
\code{\link{preprocess2}}
}
