% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREPROCESS.R
\name{PREPROCESS}
\alias{PREPROCESS}
\alias{preprocess}
\alias{PREplot}
\title{Raw data process}
\usage{
preprocess(adsl, adae)

PREplot(aedata, ptnum = 50, param = "risk difference")
}
\arguments{
\item{adsl}{subject level analysis dataset, it is a .csv file, it has to contain at least two columns, "USUBJID" and "TRTCTR", "TRTCTR"
is the indicator for treatment and control group. TRTCTR=1 for treatment group and TRTCTR=0 for control group.}

\item{adae}{adverse event analysis dataset, it is a .csv file, it has to contain at least three columns, "USUBJID", "AEBODSYS",and "AEDECOD"}

\item{aedata}{the output dataset from function \code{preprocess}}

\item{ptnum}{an integer, number of PTs to plot, with default be 50}

\item{param}{a string, summary statistics based for plotting, either 'risk difference' or 'odds ratio', with default be 'risk difference'}
}
\value{
\code{preprocess} returns a dataframe with following columnss:\cr
\emph{AEBODSYS}: Body System or SoC for AE \cr
\emph{AEDECOD}: Peferred Term (PT) for AE \cr
\emph{AEc}: number of patients in control group having each specific AE \cr
\emph{AEt}: number of patients in treatment group having each specific AE \cr
\emph{Nc}: number of total patients in control group \cr
\emph{Nt}: number of total patients in treatmetn group \cr
\emph{b}: integer represents each Soc \cr
\emph{i}: integer represents each PT \cr
\emph{j}: order of PT in each SoC \cr
\emph{Raw_Risk_Diff}: Risk difference between treatment and control group \cr
\emph{Raw_OR}: Odds ratio of the risk treatment/control
}
\description{
\code{preprocess} reorganizes the datasets by adverse events (AE), so that it can be used for further analysis.\cr
\code{PREplot} plot the risk difference or odds ratio calculated from Raw data.
}
\details{
\code{preprocess} sums up the number of patients in control and treatment group.
Also this function calculates the number of patients
experiencing the AE for each AE in control and treatment group. It also contains columns, "b", "i", "j" for further analysis.\cr
\code{PREplot} calculate the risk difference or odds ratio from Raw data and plot top AEs with highest raw risk difference or with highest
odds ratio. User can select which summary statistics('risk difference' or 'odds ratio') the plot the based on and also the number
of adverse events to plot out.
}
\note{
Make sure that adsl and adae contain the required columns with exact the same column names as listed in "parameters" section.
}
\examples{
\dontrun{
data(ADAE)
data(ADSL)
AEdata<-preprocess(adsl=ADSL, adae=ADAE)
PREplot(aedata=AEdata)
# user can use a very big number(bigger than total PTs in dataset) to plot out all the PTs
PREplot(aedata=AEdata, ptnum=50000, param='odds ratio')
}

}
